requirements.txt
Flask==3.0.0
requests==2.31.0

# ML dependencies for latent diffusion model
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
